requirements.txt
beautifulsoup4==4.6.0
dj-database-url==0.5.0
django-anymail==8.1
django-oauth-toolkit==1.3.3
django-webpack-loader==0.7.0
django==2.2.17
djangorestframework==3.12.2
html5lib==1.1
lxml==4.6.1
premailer==3.7.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python3-saml==1.10.1
pytz==2020.4
requests==2.25.0
responses==0.12.1
social-auth-app-django==3.1.0
social-auth-core==3.3.3
toolz==0.10.0

# Testing/linting tools
coverage==5.3
factory-boy==3.1.0
faker==4.17.1
freezegun==1.0.0
ipython
mypy-extensions==0.4.3
mypy==0.782
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytest-cov==2.10.1
pytest-django==3.10.0
pytest-mock==1.10.1
pytest==6.1.2

# build support
GitPython
setuptools
semver
